+++ src/test/java/net/snowflake/client/core/SessionUtilExternalBrowserTest.java
@@ -22,10 +22,15 @@ import java.net.Socket;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.nio.charset.StandardCharsets;
+import java.util.Map;
+import net.snowflake.client.AbstractDriverIT;
+import net.snowflake.client.jdbc.SnowflakeBasicDataSource;
 import net.snowflake.client.jdbc.SnowflakeSQLException;
+import net.snowflake.client.jdbc.SnowflakeSQLLoggedException;
 import net.snowflake.common.core.ClientAuthnDTO;
 import org.apache.http.client.methods.HttpPost;
 import org.apache.http.client.methods.HttpRequestBase;
+import org.junit.Ignore;
 import org.junit.Test;
 import org.mockito.MockedStatic;
 import org.mockito.Mockito;
@@ -236,5 +241,28 @@ public class SessionUtilExternalBrowserTest {
     when(loginInput.getUserName()).thenReturn("testuser");
     when(loginInput.getDisableConsoleLogin()).thenReturn(true);
     return loginInput;
+  }
+
+  // Run this test manually to confirm external browser timeout is working. When test runs it will
+  // open a browser window for authentication, close the window, and you should get the expected
+  // error message within the set timeout. Valid for driver versions after 3.18.0.
+  @Test
+  @Ignore
+  public void testExternalBrowserTimeout() throws Exception {
+    // test with username/password authentication
+    // set up DataSource object and ensure connection works
+    Map<String, String> params = AbstractDriverIT.getConnectionParameters();
+    SnowflakeBasicDataSource ds = new SnowflakeBasicDataSource();
+    ds.setServerName(params.get("host"));
+    ds.setAccount(params.get("account"));
+    ds.setPortNumber(Integer.parseInt(params.get("port")));
+    ds.setUser(params.get("user"));
+    ds.setBrowserResponseTimeout(10);
+    try {
+      ds.getConnection();
+      fail();
+    } catch (SnowflakeSQLLoggedException e) {
+      assertTrue(e.getMessage().contains("External browser authentication failed"));
+    }
   }
 }
